+++ packages/algsets/src/coll-epll/index.ts
@@ -28,10 +28,10 @@ export const collEpll: AlgSet = defineAlgSet({
       name: "AS 1",
       subset: "Anti Sune",
       algs: [
-        "y R U2 R' U' R U' R'",
+        "U R U2 R' U' R U' R'",
         "R' U' R U' R' U2 R",
-        "y2 L' U' L U' L' U2 L",
-        "y' L U2 L' U' L U' L'",
+        "U2 L' U' L U' L' U2 L",
+        "U' L U2 L' U' L U' L'",
       ],
     },
     {
@@ -39,8 +39,8 @@ export const collEpll: AlgSet = defineAlgSet({
       name: "AS 2",
       subset: "Anti Sune",
       algs: [
-        "y2 R2 D R' U R D' R' U R' U' R U' R'",
-        "y R' U' R U' R' U R' D' R U R' D R2",
+        "U2 R2 D R' U R D' R' U R' U' R U' R'",
+        "U R' U' R U' R' U R' D' R U R' D R2",
         "U2 R2 D R' U R D' R' U R' U' R U' R'",
         "R U R' f' U' R U2 R' U' R U' R' f R U' R'",
       ],
@@ -50,9 +50,9 @@ export const collEpll: AlgSet = defineAlgSet({
       name: "AS 3",
       subset: "Anti Sune",
       algs: [
-        "y2 R2 D R' U2 R D' R2 U' R U' R'",
+        "U2 R2 D R' U2 R D' R2 U' R U' R'",
         "R' U' F' R U R' U' R' F R2 U' R' U R",
-        "y2 M F' r U R' U2 r' F2 r",
+        "U2 M F' r U R' U2 r' F2 r",
         "U2 f' L F L' U2 L' U2 L U2 S",
       ],
     },
@@ -61,10 +61,10 @@ export const collEpll: AlgSet = defineAlgSet({
       name: "AS 4",
       subset: "Anti Sune",
       algs: [
-        "y2 R' U' R U' R2 D' R U2 R' D R2",
-        "y2 R U2 R' U2 r' F R F' M'",
+        "U2 R' U' R U' R2 D' R U2 R' D R2",
+        "U2 R U2 R' U2 r' F R F' M'",
         "R' U' R U R' F R U R' U' R' F' R2",
-        "y2 R U2 R' U2 L' U R U' R' L",
+        "U2 R U2 R' U2 L' U R U' R' L",
       ],
     },
     {
@@ -72,9 +72,9 @@ export const collEpll: AlgSet = defineAlgSet({
       name: "AS 5",
       subset: "Anti Sune",
       algs: [
-        "y2 r' F R F' r U R'",
+        "U2 r' F R F' r U R'",
         "R' U L U' R U L'",
-        "y2 L' U R U' L U R'",
+        "U2 L' U R U' L U R'",
         "U2 R' F R F' r U R' U' M",
       ],
     },
@@ -86,7 +86,7 @@ export const collEpll: AlgSet = defineAlgSet({
         "R U' R' U2 R U' R' U2 R' D' R U R' D R",
         "R U2 r' F R' F' r U' R U' R'",
         "R U R' F' R U2 R' U' R U' R' F R U' R'",
-        "y2 L U2 R' U L' U' R U' L U' L'",
+        "U2 L U2 R' U L' U' R U' L U' L'",
       ],
     },
     {
@@ -95,9 +95,9 @@ export const collEpll: AlgSet = defineAlgSet({
       subset: "Sune",
       algs: [
         "R U R' U R U2 R'",
-        "y' R' U2 R U R' U R",
-        "y L' U2 L U L' U L",
-        "y2 L U L' U L U2 L'",
+        "U' R' U2 R U R' U R",
+        "U L' U2 L U L' U L",
+        "U2 L U L' U L U2 L'",
       ],
     },
     {
@@ -105,7 +105,7 @@ export const collEpll: AlgSet = defineAlgSet({
       name: "S 2",
       subset: "Sune",
       algs: [
-        "y2 R U R' U R2 D R' U2 R D' R2",
+        "U2 R U R' U R2 D R' U2 R D' R2",
         "r' F2 r U2 R U' r' F M'",
         "L' U2 L U2 R U' L' U L R'",
         "L' U2 L U2 l F' L' F M'",
@@ -118,7 +118,7 @@ export const collEpll: AlgSet = defineAlgSet({
       algs: [
         "L' R U R' U' L U2 R U2 R'",
         "M F R' F' r U2 R U2 R'",
-        "y2 R2 D' R U2 R' D R2 U R' U R",
+        "U2 R2 D' R U2 R' D R2 U R' U R",
         "f R' F' R U2 R U2 R' U2 S'",
       ],
     },
@@ -127,9 +127,9 @@ export const collEpll: AlgSet = defineAlgSet({
       name: "S 4",
       subset: "Sune",
       algs: [
-        "y' R U R' U R U' R D R' U' R D' R2",
+        "U' R U R' U R U' R D R' U' R D' R2",
         "R U R' U' R' F R F' r U R' U R U2 r'",
-        "y' F R' U2 R F' R' F U2 F' R",
+        "U' F R' U2 R F' R' F U2 F' R",
         "L F' U2 F L' F' L U2 L' F",
       ],
     },
@@ -149,9 +149,9 @@ export const collEpll: AlgSet = defineAlgSet({
       name: "S 6",
       subset: "Sune",
       algs: [
-        "y2 R U R' F' R U R' U R U2 R' F R U' R'",
-        "y2 R U R' U r' F R F' r U2 R'",
-        "y2 R U R' U L' U R U' L U2 R'",
+        "U2 R U R' F' R U R' U R U2 R' F R U' R'",
+        "U2 R U R' U r' F R F' r U2 R'",
+        "U2 R U R' U L' U R U' L U2 R'",
         "F' R U2 R' U2 R' F2 R U R U' R' F'",
       ],
     },
@@ -160,9 +160,9 @@ export const collEpll: AlgSet = defineAlgSet({
       name: "L 1",
       subset: "L",
       algs: [
-        "y' R U R' U R U' R' U R U' R' U R U2 R'",
-        "y' R U2 R' U' R U R' U' R U R' U' R U' R'",
-        "y2 R' U2 R U R' U' R U R' U' R U R' U R",
+        "U' R U R' U R U' R' U R U' R' U R U2 R'",
+        "U' R U2 R' U' R U R' U' R U R' U' R U' R'",
+        "U2 R' U2 R U R' U' R U R' U' R U R' U R",
         "R' U' R U' R' U2 R U' R U R' U R U2 R'",
       ],
     },
@@ -172,9 +172,9 @@ export const collEpll: AlgSet = defineAlgSet({
       subset: "L",
       algs: [
         "R' U2 R' D' R U2 R' D R2",
-        "y2 L' U2 L' D' L U2 L' D L2",
-        "y' R' U2 R U R2 D' R U R' D R2",
-        "y' r D r' U r D' r' U y R U2 R'",
+        "U2 L' U2 L' D' L U2 L' D L2",
+        "U' R' U2 R U R2 D' R U R' D R2",
+        "U' r D r' U r D' r' U y R U2 R'",
       ],
     },
     {
@@ -182,10 +182,10 @@ export const collEpll: AlgSet = defineAlgSet({
       name: "L 3",
       subset: "L",
       algs: [
-        "y R U2 R D R' U2 R D' R2",
+        "U R U2 R D R' U2 R D' R2",
         "U2 R U2 R2 D' R U' R' D R2 U' R'",
         "R' F' R U R' U' R' F R2 U' R' U2 R",
-        "y R' U' R U2 L' U R' U' L U' R",
+        "U R' U' R U2 L' U R' U' L U' R",
       ],
     },
     {
@@ -193,10 +193,10 @@ export const collEpll: AlgSet = defineAlgSet({
       name: "L 4",
       subset: "L",
       algs: [
-        "y F R' F' r U R U' r'",
-        "y2 R2 D R' U R D' R' U' R'",
+        "U F R' F' r U R U' r'",
+        "U2 R2 D R' U R D' R' U' R'",
         "R U R' U' R' F R U R U' R' F'",
-        "y F l' U' L U R U' r'",
+        "U F l' U' L U R U' r'",
       ],
     },
     {
@@ -204,10 +204,10 @@ export const collEpll: AlgSet = defineAlgSet({
       name: "L 5",
       subset: "L",
       algs: [
-        "y2 F' r U R' U' r' F R",
-        "y x R' U R D' R' U' R D x'",
+        "U2 F' r U R' U' r' F R",
+        "U x R' U R D' R' U' R D x'",
         "r U R U' r' F R' F'",
-        "y' R2 D' R U' R' D R U R",
+        "U' R2 D' R U' R' D R U R",
       ],
     },
     {
@@ -215,10 +215,10 @@ export const collEpll: AlgSet = defineAlgSet({
       name: "L 6",
       subset: "L",
       algs: [
-        "y r U2 R2 F R F' R U2 r'",
-        "y' R' U' R U R' F' R U R' U' R' F R2",
+        "U r U2 R2 F R F' R U2 r'",
         "U' R' U' R U R' F' R U R' U' R' F R2",
-        "y F R U R2 F R F' R U' R' F'",
+        "U' R' U' R U R' F' R U R' U' R' F R2",
+        "U F R U R2 F R F' R U' R' F'",
       ],
     },
     {
@@ -227,9 +227,9 @@ export const collEpll: AlgSet = defineAlgSet({
       subset: "U",
       algs: [
         "R' U' R U' R' U2 R2 U R' U R U2 R'",
-        "y2 R U R' U R U2 R2 U' R U' R' U2 R",
-        "y R U2 R' U' R U' R' U' R U R' U R U2 R'",
-        "y' R U R' U' R U' R' U2 R U' R' U2 R U R'",
+        "U2 R U R' U R U2 R2 U' R U' R' U2 R",
+        "U R U2 R' U' R U' R' U' R U R' U R U2 R'",
+        "U' R U R' U' R U' R' U2 R U' R' U2 R U R'",
       ],
     },
     {
@@ -238,9 +238,9 @@ export const collEpll: AlgSet = defineAlgSet({
       subset: "U",
       algs: [
         "R' F R U' R' U' R U R' F' R U R' U' R' F R F' R",
-        "y' r U R' U' r' F R2 U' R' U' R U2 R' U' F'",
-        "y' R' U' R F R2 D' R U R' D R2 U' F'",
-        "y F U R U2 R' U R U R2 F' r U R U' r'",
+        "U' r U R' U' r' F R2 U' R' U' R U2 R' U' F'",
+        "U' R' U' R F R2 D' R U R' D R2 U' F'",
+        "U F U R U2 R' U R U R2 F' r U R U' r'",
       ],
     },
     {
@@ -248,7 +248,7 @@ export const collEpll: AlgSet = defineAlgSet({
       name: "U 3",
       subset: "U",
       algs: [
-        "y2 R2 D R' U2 R D' R' U2 R'",
+        "U2 R2 D R' U2 R D' R' U2 R'",
         "R U' R' U' R U2 R' U' R' D' R U2 R' D R",
         "R' U R U R' F' R U R' U' R' F R2 U' R' U' R",
         "L2 D L' U2 L D' L' U2 L'",
@@ -260,9 +260,9 @@ export const collEpll: AlgSet = defineAlgSet({
       subset: "U",
       algs: [
         "F R U' R' U R U R' U R U' R' F'",
-        "y2 R' F2 R U2 R U2 R' F2 R U2 R'",
-        "y' F U2 R' D' R U2 R' D R F'",
-        "y2 R U2 R' U2 L' U2 R U2 R' U2 L",
+        "U2 R' F2 R U2 R U2 R' F2 R U2 R'",
+        "U' F U2 R' D' R U2 R' D R F'",
+        "U2 R U2 R' U2 L' U2 R U2 R' U2 L",
       ],
     },
     {
@@ -271,7 +271,7 @@ export const collEpll: AlgSet = defineAlgSet({
       subset: "U",
       algs: [
         "R2 D' R U2 R' D R U2 R",
-        "y2 L2 D' L U2 L' D L U2 L",
+        "U2 L2 D' L U2 L' D L U2 L",
         "R2 F' R U R' U' R' F R' U' R2 U2 R2 U R' U R",
         "L U' R U' L' U R' U2 L U' L'",
       ],
@@ -293,9 +293,9 @@ export const collEpll: AlgSet = defineAlgSet({
       subset: "T",
       algs: [
         "R U2 R' U' R U' R2 U2 R U R' U R",
-        "y' R U R2 U' R2 U' R2 U2 R U' R U' R'",
+        "U' R U R2 U' R2 U' R2 U2 R U' R U' R'",
         "R U2 R' r' F2 r U' R U' R' U' r' F r",
-        "y' R U R' U R U2 R' L' U' L U' L' U2 L",
+        "U' R U R' U R U2 R' L' U' L U' L' U2 L",
       ],
     },
     {
@@ -305,8 +305,8 @@ export const collEpll: AlgSet = defineAlgSet({
       algs: [
         "R' U R U2 R' L' U R U' L",
         "R' U R U2 r' R' F R F' r",
-        "y2 R' F R U R' U' R' F' R2 U' R' U2 R",
-        "y2 R U' R' U2 L R U' R' U L'",
+        "U2 R' F R U R' U' R' F' R2 U' R' U2 R",
+        "U2 R U' R' U2 L R U' R' U L'",
       ],
     },
     {
@@ -314,10 +314,10 @@ export const collEpll: AlgSet = defineAlgSet({
       name: "T 3",
       subset: "T",
       algs: [
-        "y R' F' r U R U' r' F",
-        "y l' U' L U R U' r' F",
-        "y2 R' U' R' D' R U R' D R2",
-        "y l' U' L U l F' L' F",
+        "U R' F' r U R U' r' F",
+        "U l' U' L U R U' r' F",
+        "U2 R' U' R' D' R U R' D R2",
+        "U l' U' L U l F' L' F",
       ],
     },
     {
@@ -325,10 +325,10 @@ export const collEpll: AlgSet = defineAlgSet({
       name: "T 4",
       subset: "T",
       algs: [
-        "y2 F R U R' U' R U' R' U' R U R' F'",
-        "y2 F R' D' R U2 R' D R U2 F'",
-        "y2 R F R' U R U2 R' U R U F' R'",
-        "y R U2 R' F2 R U2 R' U2 R' F2 R",
+        "U2 F R U R' U' R U' R' U' R U R' F'",
+        "U2 F R' D' R U2 R' D R U2 F'",
+        "U2 R F R' U R U2 R' U R U F' R'",
+        "U R U2 R' F2 R U2 R' U2 R' F2 R",
       ],
     },
     {
@@ -336,9 +336,9 @@ export const collEpll: AlgSet = defineAlgSet({
       name: "T 5",
       subset: "T",
       algs: [
-        "y' r U R' U' r' F R F'",
+        "U' r U R' U' r' F R F'",
         "R U R D R' U' R D' R2",
-        "y2 x' D R U' R' D' R U R' x",
+        "U2 x' D R U' R' D' R U R' x",
         "R U R' U R' D' R U' R' D R2 U' R'",
       ],
     },
@@ -348,9 +348,9 @@ export const collEpll: AlgSet = defineAlgSet({
       subset: "T",
       algs: [
         "R' U R2 D r' U2 r D' R2 U' R",
-        "y2 R U' R2 D' r U2 r' D R2 U R'",
-        "y R' U' R U R2 D' R U2 R' D R2 U' R' U R",
-        "y R U R' U' R2 D R' U2 R D' R2 U R U' R'",
+        "U2 R U' R2 D' r U2 r' D R2 U R'",
+        "U R' U' R U R2 D' R U2 R' D R2 U' R' U R",
+        "U R U R' U' R2 D R' U2 R D' R2 U R U' R'",
       ],
     },
     {
@@ -360,7 +360,7 @@ export const collEpll: AlgSet = defineAlgSet({
       algs: [
         "R U2 R2 U' R2 U' R2 U2 R",
         "R' U2 R2 U R2 U R2 U2 R'",
-        "y2 L' U2 L2 U L2 U L2 U2 L'",
+        "U2 L' U2 L2 U L2 U L2 U2 L'",
         "R U R' U R U2 R' U' R U R' U R U2 R'",
       ],
     },
@@ -369,9 +369,9 @@ export const collEpll: AlgSet = defineAlgSet({
       name: "Pi 2",
       subset: "Pi",
       algs: [
-        "y F U R U' R' U R U' R2 F' R U R U' R'",
+        "U F U R U' R' U R U' R2 F' R U R U' R'",
         "R' F2 R U2 R U2 R' F2 U' R U' R'",
-        "y2 L' U' L U L F' L2 U' L U L' U' L U F",
+        "U2 L' U' L U L F' L2 U' L U L' U' L U F",
         "R U R' U R' F R2 U' R' U' R U R' F' U R U' R'",
       ],
     },
@@ -381,9 +381,9 @@ export const collEpll: AlgSet = defineAlgSet({
       subset: "Pi",
       algs: [
         "R' U' F' R U R' U' R' F R2 U2 R' U2 R",
-        "y F U R U' R' U R U2 R' U' R U R' F'",
-        "y F R2 U' R2 U R2 U S R2 f'",
-        "y' R U R' U R U2 R2 F' r U R U' r' F",
+        "U F U R U' R' U R U2 R' U' R U R' F'",
+        "U F R2 U' R2 U R2 U S R2 f'",
+        "U' R U R' U R U2 R2 F' r U R U' r' F",
       ],
     },
     {
@@ -393,8 +393,8 @@ export const collEpll: AlgSet = defineAlgSet({
       algs: [
         "R U R' U' R' F R2 U R' U' R U R' U' F'",
         "R U2 R' U' R U R' U2 r' F R F' M'",
-        "y F' R U R' U R U' R' U' R' F R U' R U' R' U R U R'",
-        "y' R' U2 R U R' U R2 U' L' U R' U' L",
+        "U F' R U R' U R U' R' U' R' F R U' R U' R' U R U R'",
+        "U' R' U2 R U R' U R2 U' L' U R' U' L",
       ],
     },
     {
@@ -403,9 +403,9 @@ export const collEpll: AlgSet = defineAlgSet({
       subset: "Pi",
       algs: [
         "R U' L' U R' U L U L' U L",
-        "y' R U2 R' U R' D' R U2 R' D R2 U' R'",
-        "y' R U R' U F' R U2 R' U2 R' F R",
-        "y2 L' U R U' L U' R' U' R U' R'",
+        "U' R U2 R' U R' D' R U2 R' D R2 U' R'",
+        "U' R U R' U F' R U2 R' U2 R' F R",
+        "U2 L' U R U' L U' R' U' R U' R'",
       ],
     },
     {
@@ -414,7 +414,7 @@ export const collEpll: AlgSet = defineAlgSet({
       subset: "Pi",
       algs: [
         "R' F' U' F U' R U S' R' U R S",
-        "y' r U R' U R' F R F' R U' R' U R U2 r'",
+        "U' r U R' U R' F R F' R U' R' U R U2 r'",
         "R2 D' R U R' D R U R U' R' U R U R' U R",
         "R U D' R U R' D R2 U' R' U' R2 U2 R",
       ],
@@ -425,9 +425,9 @@ export const collEpll: AlgSet = defineAlgSet({
       subset: "H",
       algs: [
         "R U R' U R U' R' U R U2 R'",
-        "y' R U2 R' U' R U R' U' R U' R'",
-        "y R U2 R' U' R U R' U' R U' R'",
-        "y' R' U2 R U R' U' R U R' U R",
+        "U' R U2 R' U' R U R' U' R U' R'",
+        "U R U2 R' U' R U R' U' R U' R'",
+        "U' R' U2 R U R' U' R U R' U R",
       ],
     },
     {
@@ -437,7 +437,7 @@ export const collEpll: AlgSet = defineAlgSet({
       algs: [
         "F R U' R' U R U2 R' U' R U R' U' F'",
         "f R2 S' U' R2 U' R2 U R2 F'",
-        "y2 f R U R' U' R F' R U R' U' R' S'",
+        "U2 f R U R' U' R F' R U R' U' R' S'",
         "f R U R' U' f' R U R' U' R' F R F'",
       ],
     },
@@ -457,10 +457,10 @@ export const collEpll: AlgSet = defineAlgSet({
       name: "H 4",
       subset: "H",
       algs: [
-        "y F R U R' U' R U R' U' R U R' U' F'",
-        "y F U R U' R' U R U' R' U R U' R' F'",
         "U F R U R' U' R U R' U' R U R' U' F'",
-        "y' F R U R' U' R U R' U' R U R' U' F'",
+        "U F U R U' R' U R U' R' U R U' R' F'",
+        "U F R U R' U' R U R' U' R U R' U' F'",
+        "U' F R U R' U' R U R' U' R U R' U' F'",
       ],
     },
   ],